Fix wake orientation
The transformation applied on the filaments in the trailing filament loops
seem to be working properly. However, rotation of the bound and first trailing
filaments were not performed.

This was realised after re-writing the rotation algorithm so that its
all performed at the end. Nonetheless, this will have little performance on
the computing time of the code as it depends primarily on the solver.