Add temporary fix for double rotor and its integration
Control points are in "global" reference frame (origin at centre of first rotor).
So the control points are distanced further apart while the solver does not
take that into account.

The temporary fix consists of making a variable cp which contains the "local"
location of the control points at the rotors. Basically, the coords of the first
rotor is copied in the locations of the second rotor (in the matrix).